feat(builder): reimplement the builder in Go
This is part of issue #3836, where the main builder scripts are
rewritten as a Go executable.

This feature adds an SSH server implementation to Go, which means we
have removed one of the major runtime interops on Builder.

This changes builder's internal port, and should not break other
things, but makes Flannel configuration easier.